fix update the deletion

+++ app/api/raffle/route.ts
@@ -129,14 +129,16 @@ export async function DELETE(req: NextRequest) {
 
     // Get the raffle id from req and make sure one was passed
     const raffle = req.nextUrl.searchParams.get("raffleId");
-    if (!raffle || raffle === "") return NextResponse.json({ error: "Missing parameters" }, { status: 400 });
+    const id = req.nextUrl.searchParams.get("id");
+
+    if (!raffle || raffle === "" || !id || id === "") return NextResponse.json({ error: "Missing parameters" }, { status: 400 });
     // Send the delete request
     const res = await fetch(`${process.env.NEXT_PUBLIC_TWITCH_URL}/channel_points/custom_rewards?broadcaster_id=${thisUser?.providerAccountId}&id=${raffle}`, {
         method: "DELETE",
         headers: { "client-id": process.env.NEXT_PUBLIC_TWITCH_API_KEY, authorization: "Bearer " + thisUser.access_token },
     });
 
-    await prisma.giveaways.update({ where: { id: raffle }, data: { twitchId: "" } });
+    await prisma.giveaways.update({ where: { twitchId: raffle, id: id }, data: { twitchId: "" } });
 
     // Return 200 if deleted successfully, otherwise pass the code
     return NextResponse.json({}, { status: res.status === 204 ? 200 : res.status });
